+++ pkg/inspector/utils.go
@@ -2,17 +2,19 @@ package inspector
 
 import (
 	"context"
-	"github.com/docker/docker/client"
+	"io"
 	"log"
+
+	"github.com/docker/docker/client"
 )
 
-func GetTarFromID(ctx context.Context, ID string) (string, error) {
+func GetTarFromID(ctx context.Context, ID string) (io.ReadCloser, error) {
 	var err error
 	// Use the inspector id from containerd or crio
 	cli, err := client.NewClientWithOpts(client.FromEnv, client.WithAPIVersionNegotiation())
 	if err != nil {
 		log.Printf("init docker environment failed: %v", err)
-		return "", err
+		return nil, err
 	}
 	c := DockerApi{
 		DCli: cli,
@@ -20,14 +22,15 @@ func GetTarFromID(ctx context.Context, ID string) (string, error) {
 
 	defer c.DCli.Close()
 
-	var tarFile string
+	var tarFile io.ReadCloser
+
 	if ctx.Value("tarType") == "image" {
 		tarFile, err = c.GetImageName(ID)
 	} else {
 		tarFile, err = c.GetContainerName(ID)
 		if err != nil {
 			log.Printf("expose inspector file error: %v", err)
-			return "", err
+			return nil, err
 		}
 
 	}