+++ pkg/packages/get_package.go
@@ -51,7 +51,7 @@ func (s *Packages) GetApp(ctx context.Context) error {
 
 	err = s.getSitePacks(ctx)
 	err = s.getNodeModulePacks(ctx)
-	err = s.getSpecialPacks(ctx)
+	err = s.Traverse(ctx)
 
 	return err
 }